backend/requirements.txt
flask
flask-cors
requests
python-dotenv
pymongo
pandas
numpy
xgboost
scikit-learn
joblib
